requirements.txt
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.3
contourpy==1.3.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.56.0
glcontext==3.0.0
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.6
moderngl==5.12.0
nest-asyncio==1.5.6
numpy==2.2.3
packaging==24.2
parso==0.8.3
pickleshare==0.7.5
pillow==11.1.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pygame==2.6.1
pyglm==2.8.0
Pygments==2.15.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==25.0.2
six==1.17.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
wcwidth==0.2.6
